requirements.txt
# Brain Caption Ground Truth Generation - Requirements
# Generated for Greene HPC environment

# Core ML/AI packages
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Image processing
Pillow>=9.0.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# HuggingFace ecosystem
huggingface-hub>=0.15.0

# COCO dataset handling
pycocotools>=2.0.4

# Data formats
h5py>=3.7.0
nibabel>=4.0.0

# Progress and visualization
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Additional utilities
requests>=2.28.0
json5>=0.9.0
